Added Animations to Main Scene
I modified some of your scripts kate! I kept pretty much all of your code but just changed references to make it a little more efficient as well as work better with the rest of the scripts!

The foxes are now drop in able, although not all the movement scripts have been made perfect for them yet, but almost there on that note. I added in the owl and set him up so he is easy for Yousab to script into.

I copied pretty much every asset from your folder, kate, and moved them to the main folder so that we keep everything under one umbrella. I want to move a few of your scripts too but Ill ask you about it in class since I don't want to mess up your scene although I think I might have as I renamed your AnimationEvents to AnimationEventsOld so I didnt have to change all of the animation controller and clip references 😧